Rename `mont` to `ecc` in circuit code.
