+++ src/circuit/mod.rs
@@ -6,7 +6,7 @@ pub mod uint32;
 pub mod blake2s;
 pub mod num;
 pub mod lookup;
-pub mod mont;
+pub mod ecc;
 pub mod pedersen_hash;
 
 use bellman::SynthesisError;
