+++ src/circuit/pedersen_hash.rs
@@ -1,5 +1,5 @@
 use super::*;
-use super::mont::{
+use super::ecc::{
     MontgomeryPoint,
     EdwardsPoint
 };
